requirements.txt
streamlit
google-generativeai 
langchain-google-genai 
python-dotenv
langchain-community
PyPDF2
pillow